Extracted all classes from main.cpp
